+++ front-end/src/pages/feedback/components/FeedbackSend3.jsx
@@ -14,7 +14,7 @@ import { transformToBytes } from '../../../utility/inputChecker';
 import {
   useFrequnetFeedbackSend,
   useRegularFeedbackSend,
-} from '../../../api/useFeedback2';
+} from '../../../api/useFeedback';
 import { showToast } from '../../../utility/handleToast';
 import { hideModal, showModal } from '../../../utility/handleModal';
 import Modal, { ModalType } from '../../../components/modals/Modal';
